backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-magic==0.4.27
Pillow==10.1.0
moviepy==1.0.3
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
opencv-python==4.8.1.78
numpy==1.24.3
slowapi==0.1.9
libmagic==1.0 